# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Global Finance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  name?: string;
  email?: string;
  profileImage?: string;
  isAdmin: boolean;
}

export interface GetCurrentAuthUserResponse {
  isAuthenticated: boolean;
  user?: AuthUser;
}

export type LoanStatus = (typeof LoanStatus)[keyof typeof LoanStatus];

export const LoanStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Loan {
  id: number;
  userId: string;
  fullName: string;
  country: string;
  phoneNumber: string;
  monthlyIncomeRange: string;
  amount: number;
  duration: number;
  status: LoanStatus;
  repaymentDate?: string;
  monthlyPayment?: number;
  createdAt?: string;
}

export type AdminLoanStatus =
  (typeof AdminLoanStatus)[keyof typeof AdminLoanStatus];

export const AdminLoanStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface AdminLoan {
  id: number;
  userId: string;
  userName?: string;
  userEmail?: string;
  fullName: string;
  country: string;
  phoneNumber: string;
  monthlyIncomeRange: string;
  amount: number;
  duration: number;
  status: AdminLoanStatus;
  repaymentDate?: string;
  monthlyPayment?: number;
  createdAt?: string;
}

export interface AdminUser {
  id: string;
  name?: string;
  email?: string;
  isAdmin: boolean;
  createdAt?: string;
}

export interface LoanApplication {
  fullName: string;
  country: string;
  phoneNumber: string;
  monthlyIncomeRange: string;
  amount: number;
  duration: number;
}

export type UpdateLoanStatusStatus =
  (typeof UpdateLoanStatusStatus)[keyof typeof UpdateLoanStatusStatus];

export const UpdateLoanStatusStatus = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateLoanStatus {
  status: UpdateLoanStatusStatus;
}

export interface ErrorResponse {
  error: string;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}