# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Global Finance API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetCurrentAuthUserResponse = zod.object({
  isAuthenticated: zod.boolean(),
  user: zod
    .object({
      id: zod.string(),
      name: zod.string().optional(),
      email: zod.string().optional(),
      profileImage: zod.string().optional(),
      isAdmin: zod.boolean(),
    })
    .optional(),
});

/**
 * @summary Get current user's loans
 */
export const GetUserLoansResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  fullName: zod.string(),
  country: zod.string(),
  phoneNumber: zod.string(),
  monthlyIncomeRange: zod.string(),
  amount: zod.number(),
  duration: zod.number(),
  status: zod.enum(["pending", "approved", "rejected"]),
  repaymentDate: zod.string().optional(),
  monthlyPayment: zod.number().optional(),
  createdAt: zod.string().optional(),
});
export const GetUserLoansResponse = zod.array(GetUserLoansResponseItem);

/**
 * @summary Apply for a loan
 */
export const ApplyForLoanBody = zod.object({
  fullName: zod.string(),
  country: zod.string(),
  phoneNumber: zod.string(),
  monthlyIncomeRange: zod.string(),
  amount: zod.number(),
  duration: zod.number(),
});

/**
 * @summary Admin - get all loan applications
 */
export const AdminGetAllLoansResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  userName: zod.string().optional(),
  userEmail: zod.string().optional(),
  fullName: zod.string(),
  country: zod.string(),
  phoneNumber: zod.string(),
  monthlyIncomeRange: zod.string(),
  amount: zod.number(),
  duration: zod.number(),
  status: zod.enum(["pending", "approved", "rejected"]),
  repaymentDate: zod.string().optional(),
  monthlyPayment: zod.number().optional(),
  createdAt: zod.string().optional(),
});
export const AdminGetAllLoansResponse = zod.array(AdminGetAllLoansResponseItem);

/**
 * @summary Admin - approve or reject a loan
 */
export const AdminUpdateLoanStatusParams = zod.object({
  loanId: zod.coerce.number(),
});

export const AdminUpdateLoanStatusBody = zod.object({
  status: zod.enum(["approved", "rejected"]),
});

export const AdminUpdateLoanStatusResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  fullName: zod.string(),
  country: zod.string(),
  phoneNumber: zod.string(),
  monthlyIncomeRange: zod.string(),
  amount: zod.number(),
  duration: zod.number(),
  status: zod.enum(["pending", "approved", "rejected"]),
  repaymentDate: zod.string().optional(),
  monthlyPayment: zod.number().optional(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Admin - get all users
 */
export const AdminGetAllUsersResponseItem = zod.object({
  id: zod.string(),
  name: zod.string().optional(),
  email: zod.string().optional(),
  isAdmin: zod.boolean(),
  createdAt: zod.string().optional(),
});
export const AdminGetAllUsersResponse = zod.array(AdminGetAllUsersResponseItem);