# lib/api-zod/src/generated/types/adminLoan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Global Finance API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminLoanStatus } from "./adminLoanStatus";

export interface AdminLoan {
  id: number;
  userId: string;
  userName?: string;
  userEmail?: string;
  fullName: string;
  country: string;
  phoneNumber: string;
  monthlyIncomeRange: string;
  amount: number;
  duration: number;
  status: AdminLoanStatus;
  repaymentDate?: string;
  monthlyPayment?: number;
  createdAt?: string;
}