# lib/api-zod/src/generated/types/exchangeMobileAuthorizationCodeResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Global Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}