# lib/api-zod/src/generated/types/loan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Global Finance API
 * OpenAPI spec version: 0.1.0
 */
import type { LoanStatus } from "./loanStatus";

export interface Loan {
  id: number;
  userId: string;
  fullName: string;
  country: string;
  phoneNumber: string;
  monthlyIncomeRange: string;
  amount: number;
  duration: number;
  status: LoanStatus;
  repaymentDate?: string;
  monthlyPayment?: number;
  createdAt?: string;
}