# lib/api-zod/src/generated/types/loanApplication.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Global Finance API
 * OpenAPI spec version: 0.1.0
 */

export interface LoanApplication {
  fullName: string;
  country: string;
  phoneNumber: string;
  monthlyIncomeRange: string;
  amount: number;
  duration: number;
}