# lib/api-zod/src/generated/types/updateLoanStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Trust Global Finance API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateLoanStatusStatus } from "./updateLoanStatusStatus";

export interface UpdateLoanStatus {
  status: UpdateLoanStatusStatus;
}